requirements.txt
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.5
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
importlib-metadata==3.7.2
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.8.6
pyserial==3.5
pytz==2020.5
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
gunicorn==20.0.4